+++ SF/include/SFSimulator.h
@@ -522,8 +522,8 @@ namespace SF
 	float getPlatformRotationXZ();
 	float getPlatformRotationYZ();
 
-	float getAgentPressure(size_t index);
-	float getObstaclePressure(size_t index);
+	double getAgentPressure(size_t index);
+	double getObstaclePressure(size_t index);
 
 	std::vector<size_t> getAgentNeighboursIndexList(size_t index, float radius);
